Fix problems with building jar and dockerization project - Make possible to run application via docker-compose - Refactor some config parts, to be more suitable
