+++ app/cache/src/main/java/com/simple/cache/config/SimpleCacheConfig.java
@@ -1,0 +1,37 @@
+package com.simple.cache.config;
+
+import java.time.Duration;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.data.redis.serializer.RedisSerializer;
+
+@Configuration
+@ComponentScan("com.simple.cache")
+public class SimpleCacheConfig {
+    @Value("${cache.ttl:PT10S}") private String ttl; // TimeToLive = 10 seconds
+
+    @Bean
+    public RedisTemplate<String, String> redisTemplate() {
+        final RedisTemplate<String, String> redisTemplate = new RedisTemplate<>();
+        redisTemplate.setKeySerializer(RedisSerializer.string());
+        redisTemplate.setValueSerializer(RedisSerializer.string());
+        redisTemplate.setConnectionFactory(redisConnectionFactory());
+        return redisTemplate;
+    }
+
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory() {
+        return new LettuceConnectionFactory();
+    }
+
+    @Bean
+    public Duration ttl() {
+        return Duration.parse(ttl);
+    }
+}
