+++ app/cache/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,1 @@
-com.simple.cache.config.SimpleRedisConfig
+com.simple.cache.config.SimpleCacheConfig
