+++ app/kafka/src/main/java/com/simple/kafka/handler/SimpleCreateRequestMessageHandlerImpl.java
@@ -9,6 +9,7 @@ import com.simple.kafka.config.SimpleKafkaTopicsProperties;
 import com.simple.models.enums.SimpleKafkaMessageType;
 import com.simple.models.requests.SimpleCreateRequest;
 import com.simple.service.api.converter.SimpleKafkaConverter;
+import com.simple.service.api.date.SimpleDateProviderService;
 import com.simple.service.api.service.SimpleService;
 
 import lombok.RequiredArgsConstructor;
@@ -21,6 +22,7 @@ public class SimpleCreateRequestMessageHandlerImpl implements SimpleMessageHandl
     private final SimpleKafkaConverter kafkaConverter;
     private final SimpleKafkaTopicsProperties kafkaTopicsProperties;
     private final SimpleService service;
+    private final SimpleDateProviderService dateProviderService;
 
     @Override
     public void handle(final byte[] payload, final Acknowledgment acknowledgment) {
