+++ docker/Dockerfile
@@ -1,5 +1,5 @@
 FROM eclipse-temurin:17
-ARG JAR_FILE=target/*.jar
-COPY ${JAR_FILE} application.jar
+ARG JAR_FILE=jar/app.jar
+COPY ${JAR_FILE} app.jar
 EXPOSE 8080
-ENTRYPOINT ["java","-jar","/application.jar"]
+ENTRYPOINT ["java","-jar","app.jar"]
