+++ docker/otel-config.yml
@@ -2,7 +2,7 @@ receivers:
   prometheus:
     config:
       scrape_configs:
-        - job_name: "example"
+        - job_name: "test-job"
           scrape_interval: 5s
           metrics_path: '/actuator/prometheus'
           static_configs:
